Reorganized interfaces in separate folder
