+++ src/console/interfaces/interface_dimensions.py
@@ -1,0 +1,17 @@
+"""Interface class for dimensions."""
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True)
+class Dimensions:
+    """Dataclass for definition of dimensional parameters."""
+
+    x: float | int  # pylint: disable=invalid-name
+    """X dimension."""
+
+    y: float | int  # pylint: disable=invalid-name
+    """Y dimension."""
+
+    z: float | int  # pylint: disable=invalid-name
+    """Z dimension."""
